+++ changelog_entry.yaml
@@ -1,4 +1,5 @@
 bump: minor
 changes:
   added:
-  - Add new immigration rules to ACA.
+  - Add new immigration rules to ACA reform.
+  - Adjust immigration formula on current ACA file.
