+++ policyengine_us/parameters/gov/aca/ineligible_immigration_statuses.yaml
@@ -3,15 +3,9 @@ values:
   2018-01-01:
     - DACA_TPS
     - UNDOCUMENTED
-  2027-01-01:
-    - ASYLEE
-    - DACA_TPS
-    - DEPORTATION_WITHHELD
-    - PAROLED_ONE_YEAR
-    - UNDOCUMENTED 
 
 metadata:
-  label: Premium Tax Credit ineligible immigration status 
+  label: Reconciliation Premium Tax Credit immigration status adjustment in effect
   period: year
   unit: list
   reference:
