+++ policyengine_us/parameters/gov/contrib/reconciliation/aca_pta_immigration_status/in_effect.yaml
@@ -1,4 +1,4 @@
-description: The Committee on Ways and Means proposed to limit the Premium Tax Credit to filers with certain immigration status , which comes in effect if this is true. 
+description: The Committee on Ways and Means proposed to limit the Premium Tax Credit to filers of certain immigration statuses , which comes in effect if this is true. 
 
 values:
   0000-01-01: false
