--- policyengine_us/reforms/aca/__init__.py
@@ -1,3 +1,0 @@
-from .aca_ptc_add_immigration_requirement import (
-    create_aca_ptc_add_immigration_requirement_reform,
-)
