+++ policyengine_us/reforms/reconciliation/__init__.py
@@ -16,3 +16,6 @@ from .ctc_ssn import (
 from .reconciled_additional_senior_standard_deduction import (
     create_reconciled_additional_senior_standard_deduction_reform,
 )
+from .aca_ptc_immigration_status import (
+    create_aca_ptc_immigration_status_reform,
+)
