+++ policyengine_us/reforms/reforms.py
@@ -108,11 +108,9 @@ from .reconciliation import (
     create_reconciled_auto_loan_interest_ald_reform,
     create_ctc_ssn_reform,
     create_reconciled_additional_senior_standard_deduction_reform,
+    create_aca_ptc_immigration_status_reform,
 )
 
-from .aca import (
-    create_aca_ptc_add_immigration_requirement_reform,
-)
 
 from policyengine_core.reforms import Reform
 import warnings
@@ -244,7 +242,9 @@ def create_structural_reforms_from_parameters(parameters, period):
             parameters, period
         )
     )
-    aca_ptc_add_immigration_requirement = create_aca_ptc_add_immigration_requirement_reform(parameters, period)
+    aca_ptc_immigration_status = create_aca_ptc_immigration_status_reform(
+        parameters, period
+    )
 
     reforms = [
         afa_reform,
@@ -295,7 +295,7 @@ def create_structural_reforms_from_parameters(parameters, period):
         reconciled_auto_loan_interest_ald,
         ctc_ssn,
         reconciled_additional_senior_standard_deduction,
-        aca_ptc_add_immigration_requirement,
+        aca_ptc_immigration_status,
     ]
     reforms = tuple(filter(lambda x: x is not None, reforms))
